Time: 0 ms (100.00%) | Memory: 16.7 MB (7.17%) - LeetSync

+++ 152-maximum-product-subarray/maximum-product-subarray.cpp
@@ -1,24 +1,26 @@
 class Solution {
 public:
     int maxProduct(vector<int>& nums) {
-        int n=nums.size();
-        int pre=1;
-        int suff=1;
-        int maxm=INT_MIN;
-        for(int i=0;i<n;i++){
-             if(pre==0){
-                pre=1;
+        int n = nums.size();
+        double pre = 1.0;
+        double suff = 1.0;
+        double maxm = INT_MIN;
+
+        for (int i = 0; i < n; i++) {
+            if (pre == 0.0) {
+                pre = 1.0;
             }
-            if(suff==0){
-                suff=1;
+            if (suff == 0.0) {
+                suff = 1.0;
             }
-            pre*=nums[i];
-            suff*=nums[n-i-1];
-           
-            maxm=max(maxm,max(pre,suff));
-            
+            pre *= nums[i];
+            suff *= nums[n - i - 1];
+
+            maxm = max(maxm, max(pre, suff));
         }
-        return maxm;
+
+        return static_cast<int>(maxm);
+        
     }
 
 };
